Split moved code files into generic and libzypp specific parts
