+++ zypp-core/Digest.cc
@@ -22,26 +22,12 @@
 #include <iostream>
 #include <sstream>
 
-#ifdef DIGEST_TESTSUITE
-#include <fstream>
-#endif
-
-#include <zypp/Digest.h>
-#include <zypp/base/PtrTypes.h>
+#include <zypp-core/Digest.h>
+#include <zypp-core/base/PtrTypes.h>
 
 using std::endl;
 
 namespace zypp {
-
-    bool DigestReport::askUserToAcceptNoDigest( const zypp::Pathname &file )
-    { return false; }
-
-    bool DigestReport::askUserToAccepUnknownDigest( const Pathname &file, const std::string &name )
-    { return false; }
-
-    bool DigestReport::askUserToAcceptWrongDigest( const Pathname &file, const std::string &requested, const std::string &found )
-    { return false; }
-
 
     const std::string & Digest::md5()
     { static std::string _type( "md5" ); return _type; }
@@ -321,39 +307,5 @@ namespace zypp {
       std::istringstream is( input );
       return digest( name, is, bufsize );
     }
-
-#ifdef DIGEST_TESTSUITE
-    int main(int argc, char *argv[])
-    {
-      bool openssl = false;
-      unsigned argpos = 1;
-
-      if(argc > 1 && std::string(argv[argpos]) == "--openssl")
-      {
-    	openssl = true;
-    	++argpos;
-      }
-
-      if(argc - argpos < 2)
-      {
-    	cerr << "Usage: " << argv[0] << " <DIGESTNAME> <FILE>" << endl;
-    	return 1;
-      }
-
-      const char* digestname = argv[argpos++];
-      const char* fn = argv[argpos++];
-
-      std::ifstream file(fn);
-
-      std::string digest = Digest::digest(digestname, file);
-
-      if(openssl)
-    	cout << digestname << "(" << fn << ")= " << digest << endl;
-      else
-    	cout << digest << "  " << fn << endl;
-
-      return 0;
-    }
-#endif
 
 } // namespace zypp
