+++ zypp/CMakeLists.txt
@@ -211,7 +211,6 @@ SET( zypp_base_SRCS
   base/Measure.cc
   base/GzStream.cc
   base/InputStream.cc
-  base/ReferenceCounted.cc
   base/SetRelationMixin.cc
   base/StrMatcher.h
   base/StrMatcher.cc
@@ -954,8 +953,6 @@ macro( ADDZYPPLIB LIBNAME )
   TARGET_LINK_LIBRARIES(${LIBNAME} ${LIBXML2_LIBRARIES} )
   TARGET_LINK_LIBRARIES(${LIBNAME} ${ZLIB_LIBRARY} )
   TARGET_LINK_LIBRARIES(${LIBNAME} ${LibSolv_LIBRARIES} )
-  TARGET_LINK_LIBRARIES(${LIBNAME} ${OPENSSL_LIBRARIES} )
-  TARGET_LINK_LIBRARIES(${LIBNAME} ${CRYPTO_LIBRARIES} )
   TARGET_LINK_LIBRARIES(${LIBNAME} ${SIGNALS_LIBRARY})
   TARGET_LINK_LIBRARIES(${LIBNAME} ${Boost_THREAD_LIBRARY})
   TARGET_LINK_LIBRARIES(${LIBNAME} ${GPGME_PTHREAD_LIBRARIES})
