+++ zypp/CheckSum.h
@@ -1,0 +1,17 @@
+/*---------------------------------------------------------------------\
+|                          ____ _   __ __ ___                          |
+|                         |__  / \ / / . \ . \                         |
+|                           / / \ V /|  _/  _/                         |
+|                          / /__ | | | | | |                           |
+|                         /_____||_| |_| |_|                           |
+|                                                                      |
+\---------------------------------------------------------------------*/
+/** \file	zypp/CheckSum.h
+ *
+*/
+#ifndef ZYPP_CHECKSUM_H
+#define ZYPP_CHECKSUM_H
+
+#include <zypp-core/CheckSum.h>
+
+#endif // ZYPP_CHECKSUM_H
