+++ zypp/base/ReferenceCounted.h
@@ -1,0 +1,17 @@
+/*---------------------------------------------------------------------\
+|                          ____ _   __ __ ___                          |
+|                         |__  / \ / / . \ . \                         |
+|                           / / \ V /|  _/  _/                         |
+|                          / /__ | | | | | |                           |
+|                         /_____||_| |_| |_|                           |
+|                                                                      |
+\---------------------------------------------------------------------*/
+/** \file	zypp/base/ReferenceCounted.h
+ *
+*/
+#ifndef ZYPP_BASE_REFERENCECOUNTED_H
+#define ZYPP_BASE_REFERENCECOUNTED_H
+
+#include <zypp-core/base/ReferenceCounted.h>
+
+#endif // ZYPP_BASE_REFERENCECOUNTED_H
